refactor: streamline animation handling in home and projects pages
- Removed redundant motion imports and logic from HomePage and Projects components.
- Introduced usePageStaggerVariants hook to centralize animation variant management based on reduced motion preferences.
- Updated contact page data structure for improved clarity and maintainability.
- Enhanced contributing guidelines with local checks and build instructions for better developer experience.
