+++ commitlint.config.mjs
@@ -1,7 +1,4 @@
 /** @type {import('@commitlint/types').UserConfig} */
 export default {
   extends: ["@commitlint/config-conventional"],
-  rules: {
-    "body-max-line-length": [2, "always", 120],
-  },
 };
