+++ src/hooks/index.ts
@@ -3,6 +3,7 @@ export { useBottomNavigation } from "./useBottomNavigation";
 export { useCache } from "./useCache";
 export { useLocalStorage } from "./useLocalStorage";
 export { useMouseTracking } from "./useMouseTracking";
+export { usePageStaggerVariants } from "./usePageStaggerVariants";
 export { useProjectFilters } from "./useProjectFilters";
 export { useScrollPosition } from "./useScrollPosition";
 export { useThemeConfig } from "./useThemeConfig";
